Rename install.py to installation.py
This should make it more evident that log messages come
from an "installation" module, rather then that we are
telling people to "install" something.
