+++ pyanaconda/installation.py
